+++ lib/src/chart/line/container.dart
@@ -13,11 +13,11 @@ import 'presenter.dart';
 class LineChartTopContainer extends ChartTopContainer {
   LineChartTopContainer({
     required ChartData chartData,
-    required ChartOptions chartOptions,
+    // todo-00-last : required ChartOptions chartOptions,
     strategy.LabelLayoutStrategy? xContainerLabelLayoutStrategy,
   }) : super(
           chartData: chartData,
-          chartOptions: chartOptions,
+    // todo-00-last : chartOptions: chartOptions,
           xContainerLabelLayoutStrategy: xContainerLabelLayoutStrategy,
         ) {
     presenterCreator = LineAndHotspotLeafCreator();